requirements.txt
streamlit
requests
pandas
fpdf
Pillow
